+++ src/common/utils/index.ts
@@ -108,9 +108,9 @@ export async function checkSfdxPlugin(pluginName: string) {
       this,
       c.yellow(
         `[dependencies] Installing sfdx plugin ${c.green(pluginName)}... \nIf is stays stuck for too long, please run ${c.green(
-          `sfdx plugins:install ${pluginName}`,
-        )})`,
-      ),
+          `sfdx plugins:install ${pluginName}`
+        )})`
+      )
     );
     const installCommand = `echo y|sfdx plugins:install ${pluginName}`;
     await execCommand(installCommand, this, { fail: true, output: false });
@@ -211,7 +211,7 @@ export async function ensureGitRepository(options: any = { init: false, clone: f
           type: "text",
           name: "value",
           message: c.cyanBright(
-            "What is the URL of your git repository ? example: https://gitlab.hardis-group.com/busalesforce/monclient/monclient-org-monitoring.git",
+            "What is the URL of your git repository ? example: https://gitlab.hardis-group.com/busalesforce/monclient/monclient-org-monitoring.git"
           ),
         });
         cloneUrl = cloneUrlPrompt.value;
@@ -252,6 +252,14 @@ export async function getCurrentGitBranch(options: any = { formatted: false }) {
     return gitBranch.replace("/", "__");
   }
   return gitBranch;
+}
+
+export async function getLatestGitCommit() {
+  if (git == null) {
+    return null;
+  }
+  const log = await git().log(["-1"]);
+  return log?.latest ?? null;
 }
 
 // Select git branch and checkout & pull if requested
@@ -333,8 +341,8 @@ export async function checkGitClean(options: any) {
     } else {
       throw new SfdxError(
         `[sfdx-hardis] Branch ${c.bold(gitStatus.current)} is not clean. You must ${c.bold(
-          "commit or reset",
-        )} the following local updates:\n${c.yellow(localUpdates)}`,
+          "commit or reset"
+        )} the following local updates:\n${c.yellow(localUpdates)}`
       );
     }
   }
@@ -388,7 +396,7 @@ export async function interactiveGitAdd(options: any = { filter: [], groups: [] 
         type: "multiselect",
         name: "files",
         message: c.cyanBright(
-          `Please select ${c.red("carefully")} the ${c.bgWhite(c.red(c.bold(group.label.toUpperCase())))} files you want to commit (save)}`,
+          `Please select ${c.red("carefully")} the ${c.bgWhite(c.red(c.bold(group.label.toUpperCase())))} files you want to commit (save)}`
         ),
         choices: matchingFiles.map((fileStatus: FileStatusResult) => {
           return {
@@ -405,12 +413,12 @@ export async function interactiveGitAdd(options: any = { filter: [], groups: [] 
       result.added.push(
         ...selectFilesStatus.files
           .filter((fileStatus: FileStatusResult) => fileStatus.working_dir !== "D")
-          .map((fileStatus: FileStatusResult) => fileStatus.path.replace('"', "")),
+          .map((fileStatus: FileStatusResult) => fileStatus.path.replace('"', ""))
       );
       result.removed.push(
         ...selectFilesStatus.files
           .filter((fileStatus: FileStatusResult) => fileStatus.working_dir === "D")
-          .map((fileStatus: FileStatusResult) => fileStatus.path.replace('"', "")),
+          .map((fileStatus: FileStatusResult) => fileStatus.path.replace('"', ""))
       );
     }
     if (filesFiltered.length > 0) {
@@ -422,8 +430,8 @@ export async function interactiveGitAdd(options: any = { filter: [], groups: [] 
               .map((fileStatus: FileStatusResult) => {
                 return `  - (${getGitWorkingDirLabel(fileStatus.working_dir)}) ${getSfdxFileLabel(fileStatus.path)}`;
               })
-              .join("\n"),
-        ),
+              .join("\n")
+        )
       );
     }
     // Ask user for confirmation
@@ -484,7 +492,7 @@ export async function gitAddCommitPush(
     pattern: "./*",
     commitMessage: "Updated by sfdx-hardis",
     branch: null,
-  },
+  }
 ) {
   if (git == null) {
     if (options.init) {
@@ -523,7 +531,7 @@ export async function execSfdxJson(
     fail: false,
     output: false,
     debug: false,
-  },
+  }
 ): Promise<any> {
   if (!command.includes("--json")) {
     command += " --json";
@@ -540,7 +548,7 @@ export async function execCommand(
     output: false,
     debug: false,
     spinner: true,
-  },
+  }
 ): Promise<any> {
   let commandLog = `[sfdx-hardis][command] ${c.bold(c.bgWhite(c.grey(command)))}`;
   const execOptions: any = { maxBuffer: 10000 * 10000 };
@@ -628,7 +636,7 @@ export async function execCommand(
     return {
       status: 1,
       errorMessage: c.red(
-        `[sfdx-hardis][ERROR] Error parsing JSON in command result: ${e.message}\n${commandResult.stdout}\n${commandResult.stderr})`,
+        `[sfdx-hardis][ERROR] Error parsing JSON in command result: ${e.message}\n${commandResult.stdout}\n${commandResult.stderr})`
       ),
     };
   }
@@ -707,7 +715,7 @@ export async function filterPackageXml(
     removeStandard: false,
     removeFromPackageXmlFile: null,
     updateApiVersion: null,
-  },
+  }
 ): Promise<{ updated: boolean; message: string }> {
   let updated = false;
   let message = `[sfdx-hardis] ${packageXmlFileOut} not updated`;
@@ -796,7 +804,7 @@ export async function filterPackageXml(
 
   // Remove metadata types (named, and empty ones)
   manifest.Package.types = manifest.Package.types.filter(
-    (type: any) => !(options.removeMetadatas || []).includes(type.name[0]) && (type?.members?.length || 0) > 0,
+    (type: any) => !(options.removeMetadatas || []).includes(type.name[0]) && (type?.members?.length || 0) > 0
   );
   const builder = new xml2js.Builder({ renderOpts: { pretty: true, indent: "  ", newline: "\n" } });
   const updatedFileContent = builder.buildObject(manifest);
@@ -866,7 +874,7 @@ export async function generateReports(
   resultSorted: any[],
   columns: any[],
   commandThis: any,
-  options: any = { logFileName: null, logLabel: "Generated report files:" },
+  options: any = { logFileName: null, logLabel: "Generated report files:" }
 ): Promise<any[]> {
   const logLabel = options.logLabel || "Generated report files:";
   let logFileName = options.logFileName || null;
@@ -990,16 +998,16 @@ export async function generateSSLCertificate(branchName: string, folder: string,
     uxLog(
       commandThis,
       c.cyanBright(
-        `You must configure CI variable ${c.green(c.bold(`SFDX_CLIENT_ID_${branchName.toUpperCase()}`))} with value ${c.bold(c.green(consumerKey))}`,
-      ),
+        `You must configure CI variable ${c.green(c.bold(`SFDX_CLIENT_ID_${branchName.toUpperCase()}`))} with value ${c.bold(c.green(consumerKey))}`
+      )
     );
     uxLog(
       commandThis,
       c.cyanBright(
         `You must configure CI variable ${c.green(c.bold(`SFDX_CLIENT_KEY_${branchName.toUpperCase()}`))} with value ${c.bold(
-          c.green(encryptionKey),
-        )}`,
-      ),
+          c.green(encryptionKey)
+        )}`
+      )
     );
     uxLog(commandThis, c.yellow("Help to configure CI variables are here: https://sfdx-hardis.cloudity.com/salesforce-ci-cd-setup-auth/"));
     await prompts({
@@ -1078,16 +1086,14 @@ export async function generateSSLCertificate(branchName: string, folder: string,
     } catch (e) {
       uxLog(
         commandThis,
-        c.red("Error pushing ConnectedApp metadata. Maybe the app name is already taken ?\nYou may try again with another connected app name"),
+        c.red("Error pushing ConnectedApp metadata. Maybe the app name is already taken ?\nYou may try again with another connected app name")
       );
       uxLog(
         commandThis,
         c.yellow(`If this is a Test class issue (production env), you may have to create manually connected app ${promptResponses.appName}:
 - Follow instructions here: https://developer.salesforce.com/docs/atlas.en-us.sfdx_dev.meta/sfdx_dev/sfdx_dev_auth_connected_app.htm
   - Use certificate ${c.bold(crtFile)}
-- Once created, update variable ${c.green(
-          c.bold(`SFDX_CLIENT_ID_${branchName.toUpperCase()}`),
-        )} with ConsumerKey of the newly created connected app`),
+- Once created, update variable ${c.green(c.bold(`SFDX_CLIENT_ID_${branchName.toUpperCase()}`))} with ConsumerKey of the newly created connected app`)
       );
     }
   } else {
@@ -1096,18 +1102,16 @@ export async function generateSSLCertificate(branchName: string, folder: string,
     uxLog(
       commandThis,
       `Follow instructions here: ${c.bold(
-        "https://developer.salesforce.com/docs/atlas.en-us.sfdx_dev.meta/sfdx_dev/sfdx_dev_auth_connected_app.htm",
-      )}`,
+        "https://developer.salesforce.com/docs/atlas.en-us.sfdx_dev.meta/sfdx_dev/sfdx_dev_auth_connected_app.htm"
+      )}`
     );
     uxLog(
       commandThis,
-      `Use ${c.green(crtFile)} as certificate on Connected App configuration page, ${c.bold(`then delete ${crtFile} for security`)}`,
+      `Use ${c.green(crtFile)} as certificate on Connected App configuration page, ${c.bold(`then delete ${crtFile} for security`)}`
     );
     uxLog(
       commandThis,
-      `- configure CI variable ${c.green(
-        `SFDX_CLIENT_ID_${branchName.toUpperCase()}`,
-      )} with value of ConsumerKey on Connected App configuration page`,
+      `- configure CI variable ${c.green(`SFDX_CLIENT_ID_${branchName.toUpperCase()}`)} with value of ConsumerKey on Connected App configuration page`
     );
     uxLog(commandThis, `- configure CI variable ${c.green(`SFDX_CLIENT_KEY_${branchName.toUpperCase()}`)} with value ${c.green(encryptionKey)} key`);
   }